requirements.txt
certifi==2022.12.7
chardet==3.0.4
charset-normalizer==3.1.0
googletrans==3.1.0a0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==3.4
mysql-connector==2.2.9
oauthlib==3.2.2
PySocks==1.7.1
python-dotenv==1.0.0
requests==2.28.2
requests-oauthlib==1.3.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
tweepy==3.10.0
urllib3==1.26.15
